separate card front view into own file
