+++ CosmoPic/Views/TodayViews/TodayPhotoView.swift
@@ -64,53 +64,6 @@ struct TodayPhotoView: View {
   .environmentObject(FavoritesViewModel())
 }
 
-struct TodayPhotoCardFrontView: View {
-  let photo: Photo
-  let photoURL: URL?
-  let height: CGFloat
-  let width: CGFloat
-  let onInfoButtonTapped: () -> Void
-
-  var body: some View {
-    ZStack(alignment: .topTrailing) {
-      ZStack(alignment: .bottomTrailing) {
-        ZStack(alignment: .bottomLeading) {
-          AsyncImage(url: photoURL) { image in
-            image
-              .resizable()
-              .aspectRatio(contentMode: .fill)
-          } placeholder: {
-            HStack {
-              Spacer()
-              ProgressView()
-                .frame(height: 300)
-              Spacer()
-            }
-          }
-          .frame(width: width, height: height)
-          .cornerRadius(20)
-          .clipped()
-
-          InformationButtonView(photo: photo) {
-            withAnimation {
-              onInfoButtonTapped()
-            }
-          }
-        }
-
-        SaveButtonView(photoURL: photoURL)
-      }
-
-      VStack {
-        FavoriteButtonView(photo: photo)
-        if let photoURL {
-          ShareButtonView(photoURL: photoURL)
-        }
-      }
-    }
-  }
-}
-
 struct TodayPhotoCardBackView: View {
   let photoExplanation: String
   let height: CGFloat
